requirements.txt
requests
pillow
pytesseract
beautifulsoup4
selenium